Warning icon moved to svg folder
