+++ src/assets/svg/WarningIcon.jsx
